requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
beautifulsoup4==4.13.4
certifi==2025.4.26
charset-normalizer==3.4.2
chromadb==0.4.24
click==8.2.1
colorama==0.4.6
dataclasses-json==0.6.7
Deprecated==1.2.18
dirtyjson==1.0.8
distro==1.9.0
fastapi==0.115.12
frozenlist==1.6.0
fsspec==2025.5.0
greenlet==3.2.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.10.0
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-core==0.3.60
langchain-openai==0.3.17
langchain-text-splitters==0.3.8
langchain-chroma==0.1.1
langsmith==0.3.42
marshmallow==3.26.1
multidict==6.4.4
mypy_extensions==1.1.0
nest-asyncio==1.6.0
networkx==3.4.2
nltk==3.9.1
numpy==1.26.4
openai==1.81.0
orjson==3.10.18
packaging==23.2
pandas==2.2.3
propcache==0.3.1
pydantic==2.11.4
pydantic_core==2.33.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.31.0
requests-toolbelt==1.0.0
schedule==1.2.2
sentence-transformers==2.5.1
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.41
starlette==0.46.2
tenacity==8.5.0
tiktoken==0.9.0
tqdm==4.67.1
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==1.26.20
uvicorn==0.27.0
wrapt==1.17.2
yarl==1.20.0
zstandard==0.23.0